+++ src/c_api.cpp
@@ -13,7 +13,7 @@
 #include <LightGBM/objective_function.h>
 #include <LightGBM/prediction_early_stop.h>
 #include <LightGBM/utils/common.h>
-#include <LightGBM/utils/LocaleContext.h>
+#include <LightGBM/utils/locale_context.h>
 #include <LightGBM/utils/log.h>
 #include <LightGBM/utils/openmp_wrapper.h>
 #include <LightGBM/utils/random.h>
